+++ R/vol.R
@@ -9,7 +9,6 @@
 #' @param fill logical. if \code{TRUE} (default), fills the volume defined by the points.
 #' @param new logical. Should a new plot be started or draw over an open plot?
 #' (defaults to FALSE)
-#' @param view,scale.y,axis deprecated arguments.
 #' @param ... additional graphical options. See \code{\link{polygon}} and \code{\link{tetraplot}}.
 #'
 #' @return \code{vol} creates a 3D convex hull within a static tetrahedral plot.
@@ -24,17 +23,7 @@
 #'
 
 vol <- function(tcsdata, alpha = 0.2, grid = TRUE, fill = TRUE,
-                new = FALSE, view, scale.y, axis, ...) {
-  if (!missing(view)) {
-    stop('argument "view" is deprecated, please use "theta" and "phi" instead. see ?plot.colspace or ?tetraplot for more information.', call. = FALSE)
-  }
-  if (!missing(scale.y)) {
-    stop('argument "scale.y" is deprecated, please use "expand" instead. see ?plot.colspace or ?tetraplot for more information.', call. = FALSE)
-  }
-  if (!missing(axis)) {
-    stop('argument "axis" is deprecated, please use "box" instead. see ?plot.colspace or ?tetraplot for more information.', call. = FALSE)
-  }
-
+                new = FALSE, ...) {
 
   if (!is.null(attr(tcsdata, "clrsp")) && attr(tcsdata, "clrsp") != "tcs") stop("object is not in tetrahedral color space")